+++ nylas/client/restful_model_collection.py
@@ -63,6 +63,19 @@ class RestfulModelCollection(object):
             limit = self.filters["limit"]
         return self._range(self.filters["offset"], limit)
 
+    def count(self):
+        """
+        Get the number of objects in the collection being queried
+
+        Returns:
+            int: The number of objects in the collection being queried
+        """
+        self.filters["view"] = "count"
+        response = self.api._get_resource_raw(
+            self.model_class, resource_id=None, **self.filters
+        ).json()
+        return response["count"]
+
     def where(self, filter=None, **filters):
         # Some API parameters like "from" and "in" also are
         # Python reserved keywords. To work around this, we rename